twister: Add option to not include paths in test suites names
Test suites names are not being handled uniformly for tests not
in zephyr tree. Their names depend on -T arg used in twister's CLI.
The newly added option make twister not adding paths to suite names.
This is needed if test plans are to be used for tests outside of
zephyr tree.

Signed-off-by: Maciej Perkowski <Maciej.Perkowski@nordicsemi.no>